Task_21 - Task_40

+++ src/tasks/Task_23.java
@@ -3,9 +3,9 @@ package tasks;
 public class Task_23 {
 
     public static void main(String[] args) {
-        int a = 20;
-        int b = 1;
-        int c = 50;
+        double a = 20;
+        double b = 1;
+        double c = 50;
 
         if (a == 1 || b == 1 || c == 1) {
             System.out.println(true);
